requirements.txt
opencv-python
numpy
pytesseract
pandas
pdf2image
pillow
openpyxl
pyinstaller
